refactor: 입력과 출력 구분해서 작성
입력과 출력 구분해서 작성
