+++ java-playground/src/main/java/org/gonza/javaplayground/OutputView.java
@@ -1,8 +1,10 @@
 package org.gonza.javaplayground;
 
-import java.util.Scanner;
-
 public class OutputView {
+
+    public static void enterNumber() {
+        System.out.print("숫자를 입력해 주세요 : ");
+    }
 
     public static void strikeBallView(int strikeCount, int ballCount) {
         if (strikeCount > 0 && ballCount > 0) {
@@ -16,26 +18,12 @@ public class OutputView {
         System.out.println(ballCount + "볼");
     }
 
-    public static int gameResultView() {
+    public static void printSelectRetry() {
         System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
         System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
-        Scanner scanner = new Scanner(System.in);
-
-        int isContinue = scanner.nextInt();
-
-        try {
-            validate(isContinue);
-        } catch (Exception e) {
-            System.out.println("번호는 1, 2 중에서 입력하셔야 합니다");
-            return gameResultView();
-        }
-
-        return isContinue;
     }
 
-    private static void validate(int isContinueNumber) {
-        if (isContinueNumber != 1 && isContinueNumber != 2) {
-            throw new IllegalArgumentException();
-        }
+    public static void announceRetryException() {
+        System.out.println("번호는 1, 2 중에서 입력하셔야 합니다");
     }
 }
